app/requirements.txt
Flask
flask_sqlalchemy
requests
six
Flask-RESTful
psycopg2-binary
PyMySQL
flask_cors
shapely
pyproj